+++ week2/project/ecommerce/src/pages/productsList/ProductsList.jsx
@@ -1,7 +1,7 @@
 import Categories from './categories/Categories'
 import Products from './products/Products'
-import getCategories from '../../data/categories'
-import getProducts from '../../data/products'
+import getCategories from '../../handlers/categories'
+import getProducts from '../../handlers/products'
 import { useState, useEffect } from 'react'
 import ErrorMessage from '../../components/ErrorMessage'
 import Loading from '../../components/Loading'
